refactor: 抽离 ReviewCommandHandler（E3 Step 1a）
$v 群审核命令（handleReview/handleReviewList/handleReviewDecision/playerNameOf/relativeTime/
emitReviewUsage）抽到 ReviewCommandHandler，通过 Supplier 注入 reviewService/rankService
（调用时读取最新值，规避 setter 二阶段注入的陈旧引用）。test + integrationTest 全绿。
